new: kitty is now installable
However only for systems with /usr/bin and what not. So not Windows.

Besides that it uses GNU make and I highly doubt any Windows user has
that installed unless they're very often working in WSL or something
similar. But to be fair I'm not sure I'll ever support it.

The path at which it copies to can be changed in the Makefile, simply
modify the BIN variable.