new: standalone executable with pkg and other
I added a Makefile target called "compile" which uses a local install of
pkg to build a Windows, Linux and macOS executable that doesn't require
NodeJS, only to build it ofc. This makes it very portable. If you want
you can even build to BSD or ARM and so on...

I also made it so "install" runs "npm i" before doing anything, as that
is a requirement anyway. Same with "compile"

I also removed the /usr/bin/kitty-src as a dependency, idk how it got in
there to begin with...