new: install script for Linux and macOS (for now)
I finally added a proper install script that you can download for the
compiled version of kitty, that is both for the direct download on the
GitHub, but also when compiling.

Simply running it will detect what system you're on, then check what
folders are available, and selects the best one. Then copies the right
binary to that folder.

I'll support BSD and alike soon...