new: regEx match based file listings
This will allow one to set it up so, only when a regEx pattern matches
with the path and only then will you be able to see the file listing.
Contrary to the before binary, yes or no.

Simply by adding something like this, into the config:

    "no_filelistings": [
        ".*/test.*"
    ]

And now only folders that match that pattern will have file listings
enabled, it should be noted the path that it checks against is the full
path, that is, it is the path that you specified in the config/cli.

Aka, if you do:

    $ kitty /home/user/website

If you then navigate to the website it'll check against the whole path
aka "/home/user/website", hence why the ".*/" is needed. I might add an
option later down the line that disables this and only checks against
the requested path instead of the full path, request path being what the
browser requests not what the server looks for.

(i.e browser requests: /, server looks for /home/user/website/)

You can still just set it to true or false, and either will still work,
but if you need a more specific path to be blocked you can now do that.