implemented nearly all settings
The only settings missing are authentication and htpasswd, for the same
reasons, I also took the liberty to rename noerrorpage to no_errorpage
to make it more readable and for it to look like all the other settings
as well, as they all were using snake case as well.

I also disabled the ability to use booleans, and now only arrays are
acceptable, however not providing a value is the same as false, and
putting something like '[".*"]' is the same as true. It's just for more
flexibility and better codebase.

authentication will also be turning into an array compared to it's only
boolean value before, meaning you can require authentication on some
requests but not all.